feat(remix): Better server error handling
