--- packages/remix/src/client/utils.ts
@@ -1,19 +1,0 @@
-import { invalidClerkStatePropError, noClerkStateError } from '../errors';
-import { ClerkState } from './types';
-
-export function warnForSsr(val: ClerkState | undefined) {
-  if (!val || !val.__internal_clerk_state) {
-    console.warn(noClerkStateError);
-  }
-}
-
-export function assertValidClerkState(
-  val: any,
-): asserts val is ClerkState | undefined {
-  if (!val) {
-    throw new Error(noClerkStateError);
-  }
-  if (!!val && !val.__internal_clerk_state) {
-    throw new Error(invalidClerkStatePropError);
-  }
-}
